requirements.txt
cryptography
pycryptodome
pip install flask
pip install pycryptodome --force-reinstall
